Plyr: allow controls to become hidden when video is paused
E.g. if video is paused, and the user mouses out of the video
or keeps the mouse still for 2 seconds, the controls need to be
hidden. This is useful if the controls are blocking something
such as brief text at the bottom of the video that requires
pausing to see.

This is also in line with the behavior of the default Firefox
player.

Unfortunately, Plyr had to be changed. There was no way to monkey
patch it because the relevant function is in an inaccessible
closure.

Relevant Plyr issue: https://github.com/sampotts/plyr/issues/1251